refactor: moved sg dataval class to relevant package
